+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/key/component/AIAgentListComponent.kt
@@ -3,12 +3,21 @@ package ai.mcpdirect.studio.app.key.component
 import ai.mcpdirect.mcpdirectstudioapp.JSON
 import ai.mcpdirect.studio.app.compose.StudioActionBar
 import ai.mcpdirect.studio.app.compose.StudioListItem
+import ai.mcpdirect.studio.app.compose.StudioSearchbar
 import ai.mcpdirect.studio.app.tips.AIAgent
 import ai.mcpdirect.studio.app.tips.aiAgentIntegrationGuide
+import androidx.compose.foundation.VerticalScrollbar
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.rememberScrollbarAdapter
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -17,7 +26,9 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 
 @Composable
 fun AIAgentListComponent(
@@ -26,6 +37,8 @@ fun AIAgentListComponent(
 ){
     var aiAgents by remember {mutableStateOf<List<AIAgent>>(emptyList())}
     var aiAgent by remember { mutableStateOf<AIAgent?>(null) }
+    var filter by remember { mutableStateOf("") }
+    val listState = rememberLazyListState()
     LaunchedEffect(Unit) {
         aiAgents = JSON.decodeFromString(aiAgentIntegrationGuide)
         if(aiAgents.isNotEmpty()){
@@ -36,21 +49,33 @@ fun AIAgentListComponent(
     }
     Column(modifier) {
         StudioActionBar("AI Agents")
-        HorizontalDivider()
-        LazyColumn(Modifier.weight(1f)) {
-            items(aiAgents) {
+//        HorizontalDivider()
+        StudioSearchbar(modifier = Modifier.padding(start=16.dp, bottom = 16.dp, end = 16.dp)) {
+            filter = it
+        }
+        Box(Modifier.weight(1f)){
+            LazyColumn(state=listState) {
+                items(aiAgents) { agent->
 //                if(aiAgent==null){
 //                    aiAgent = it
 //                }
-                StudioListItem(
-                    modifier = Modifier.clickable {
-                        aiAgent = it
-                        onAIAgentChange(it)
-                    },
-                    selected = it == aiAgent,
-                    headlineContent = { Text(it.name) }
-                )
+                    if(agent.name.contains(filter,true))StudioListItem(
+                        modifier = Modifier.clickable {
+                            aiAgent = agent
+                            onAIAgentChange(agent)
+                        },
+                        selected = agent == aiAgent,
+                        headlineContent = { Text(agent.name) }
+                    )
+                }
             }
+            VerticalScrollbar(
+                modifier = Modifier
+                    .align(Alignment.CenterEnd)
+                    .fillMaxHeight(),
+                adapter = rememberScrollbarAdapter(listState)
+            )
         }
+
     }
 }